+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Outlook.Grid/UI/LanguageSettingsManager.cs
@@ -8,5 +8,16 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
 {
     public class LanguageSettingsManager : KryptonForm
     {
+        private void InitializeComponent()
+        {
+            this.SuspendLayout();
+            // 
+            // LanguageSettingsManager
+            // 
+            this.ClientSize = new System.Drawing.Size(726, 516);
+            this.Name = "LanguageSettingsManager";
+            this.ResumeLayout(false);
+
+        }
     }
 }