+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.TaskDialogs/Globals/GlobalDeclarations.cs
@@ -1,0 +1,23 @@
+﻿#region BSD License
+/*
+ * Use of this source code is governed by a BSD-style
+ * license or other governing licenses that can be found in the LICENSE.md file or at
+ * https://raw.githubusercontent.com/Krypton-Suite/Extended-Toolkit/master/LICENSE
+ */
+#endregion
+
+global using System;
+global using System.ComponentModel;
+global using System.Diagnostics.CodeAnalysis;
+global using System.Drawing;
+global using System.Drawing.Drawing2D;
+global using System.Drawing.Design;
+global using System.Drawing.Imaging;
+global using System.Media;
+global using System.Reflection;
+global using System.Runtime.InteropServices;
+global using System.Runtime.Versioning;
+global using System.Text;
+global using System.Windows.Forms;
+
+global using Krypton.Toolkit.Suite.Extended.TaskDialogs.Properties;